feat: put deconstruciton code into useFilemanagement intead of Viewer
